+++ app/Http/Controllers/Api/Admin/Subscriber/SubscriberController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\Admin;
+namespace App\Http\Controllers\Api\Admin\Subscriber;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Subscriber\SendSubscriberFormRequest;
